WP7 Sampe Works!
WP7 Smaple uses the Apgiee.NET Portable Library +  "Apgiee.WP7.CS" which
is the Windows Phone 7 implementation of the IHttpTools Interface.

The Portable Library makes unfied code where possible. the Http requests
are made differently on different platforms. this is why when you create
an Instance of ApigeeClient from the PortLib, you must provide and
Implementation for the HttpTools Interafce for thr specific Platrom.

(For now we have .NET 4.5 & WP7 Implementations available in samlpes...)